WIP: Extract `ContextEditor` sans `AssistantPanel` references
