+++ crates/assistant/src/assistant.rs
@@ -334,24 +334,6 @@ fn update_slash_commands_from_settings(cx: &mut AppContext) {
     }
 }
 
-pub fn humanize_token_count(count: usize) -> String {
-    match count {
-        0..=999 => count.to_string(),
-        1000..=9999 => {
-            let thousands = count / 1000;
-            let hundreds = (count % 1000 + 50) / 100;
-            if hundreds == 0 {
-                format!("{}k", thousands)
-            } else if hundreds == 10 {
-                format!("{}k", thousands + 1)
-            } else {
-                format!("{}.{}k", thousands, hundreds)
-            }
-        }
-        _ => format!("{}k", (count + 500) / 1000),
-    }
-}
-
 #[cfg(test)]
 #[ctor::ctor]
 fn init_logger() {
