+++ crates/assistant_context_editor/Cargo.toml
@@ -13,6 +13,7 @@ path = "src/assistant_context_editor.rs"
 
 [dependencies]
 anyhow.workspace = true
+assistant_settings.workspace = true
 assistant_slash_command.workspace = true
 assistant_slash_commands.workspace = true
 assistant_tool.workspace = true
@@ -27,32 +28,41 @@ fs.workspace = true
 futures.workspace = true
 fuzzy.workspace = true
 gpui.workspace = true
+indexed_docs.workspace = true
 language.workspace = true
 language_model.workspace = true
+language_model_selector.workspace = true
 language_models.workspace = true
 log.workspace = true
+multi_buffer.workspace = true
 open_ai.workspace = true
+parking_lot.workspace = true
 paths.workspace = true
+picker.workspace = true
 project.workspace = true
 prompt_library.workspace = true
 regex.workspace = true
+rope.workspace = true
 rpc.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+settings.workspace = true
 smallvec.workspace = true
 smol.workspace = true
 strum.workspace = true
 telemetry_events.workspace = true
 text.workspace = true
+theme.workspace = true
 ui.workspace = true
 util.workspace = true
 uuid.workspace = true
+workspace.workspace = true
 
 [dev-dependencies]
 language_model = { workspace = true, features = ["test-support"] }
-parking_lot.workspace = true
+languages = { workspace = true, features = ["test-support"] }
 pretty_assertions.workspace = true
 rand.workspace = true
-settings.workspace = true
+tree-sitter-md.workspace = true
 unindent.workspace = true
 workspace = { workspace = true, features = ["test-support"] }
