+++ crates/assistant_context_editor/src/slash_command.rs
