+++ crates/assistant_context_editor/src/slash_command_picker.rs
